341561 Exception when adding o.e.j.s.DoSFilter as managed attribute
git-svn-id: svn+ssh://dev.eclipse.org/svnroot/rt/org.eclipse.jetty/jetty/trunk@2947 7e9141cc-0065-0410-87d8-b60c137991c4

+++ jetty-servlets/src/main/java/org/eclipse/jetty/servlets/DoSFilter.java
@@ -718,7 +718,7 @@ public class DoSFilter implements Filter
      * 
      * @return number of requests
      */
-    public long getThrottledRequests()
+    public int getThrottledRequests()
     {
         return _throttledRequests;
     }
